+++ public/scripts/src/engine/thinkers/AbstractThinker.js
