+++ public/scripts/src/engine/thinkers/PlayerThinker.js
@@ -78,7 +78,6 @@ class PlayerThinker extends FPSThinker {
 
 		mob.xOfs = Math.abs(mob.xOfs) < 1 ? 0 : mob.xOfs / 2;
 		mob.yOfs = Math.abs(mob.yOfs) < 1 ? 0 : mob.yOfs / 2;
-
 		this._game.netUpdatePlayerMobile(f, x, y, sx, sy, c);
 	}
 }
