+++ public/scripts/src/engine/thinkers/index.js
