+++ public/scripts/src/store/plugins.js
@@ -1,11 +1,11 @@
 import login from '../wsplugins/login';
 import ms from '../wsplugins/ms';
-import game from '../wsplugins/game';
+import engine from '../wsplugins/engine';
 
 const socket = io(window.location.protocol + '//' + window.location.host);
 
 export default [
 	login(socket),
 	ms(socket),
-	game(socket)
+	engine(socket)
 ];
